Add Some Custom Exceptions

+++ PontiApp.EventPlace.Services/UserServices/UserService.cs
@@ -14,6 +14,7 @@ using PontiApp.Models.Request;
 using PontiApp.Models.Response;
 using PontiApp.GraphAPICalls;
 using PontiApp.AuthService;
+using PontiApp.Exceptions;
 
 namespace PontiApp.EventPlace.Services.UserServices
 {
@@ -69,22 +70,32 @@ namespace PontiApp.EventPlace.Services.UserServices
         public async Task Delete(long id)
         {
             var user = _userRepository.GetByPredicate(user => user.FbKey == id).FirstOrDefault();
+            if (user == null)
+                throw new DoesNotExistsException();
             await _userRepository.Delete(user);
         }
 
         public async Task<UserResponse> Get(long FbId)
         {
-            return _mapper.Map<UserResponse>(_userRepository.GetByPredicate(user => user.FbKey == FbId).FirstOrDefault());
+            var user = _userRepository.GetByPredicate(user => user.FbKey == FbId).FirstOrDefault();
+            if (user == null)
+                throw new DoesNotExistsException("Such User Does Not Exists!");
+            return _mapper.Map<UserResponse>(user);
         }
 
         public async Task<List<UserResponse>> GetAllUser()
         {
-            return _mapper.Map<List<UserResponse>>(await _userRepository.GetAll());
+            var allUsers = await _userRepository.GetAll();
+            if (allUsers.Count == 0 || allUsers == null)
+                throw new DoesNotExistsException("No User Exists!");
+            return _mapper.Map<List<UserResponse>>(allUsers);
         }
 
         public async Task Update(UserRequest userRequest)
         {
-            var user = _userRepository.GetByPredicate(user => user.FbKey == userRequest.fbId).First(); 
+            var user = _userRepository.GetByPredicate(user => user.FbKey == userRequest.fbId).First();
+            if (user == null)
+                throw new DoesNotExistsException("Such User Does Not Exists!");
             user.Address = userRequest.Address;
             user.AverageRanking = userRequest.AverageRanking;
             user.Mail = userRequest.Mail;
@@ -97,10 +108,20 @@ namespace PontiApp.EventPlace.Services.UserServices
 
         public bool UserExists(long FbKey) => _userRepository.GetByPredicate(user => user.FbKey == FbKey).Any();
 
-        public async Task<bool> UserExists(int id) => await _userRepository.GetByID(id) is null;
+        public async Task<bool> UserExists(int id) => await _userRepository.GetByID(id) is not null;
 
-        public async Task<UserEntity> GetUser(int id) => await _userRepository.GetByID(id);
-        public UserCreationDTO GetUser(long fbId) => _mapper.Map<UserCreationDTO>(_userRepository.GetByPredicate(f => f.FbKey == fbId).FirstOrDefault());
+        public async Task<UserEntity> GetUser(int id) {
+            var user = await _userRepository.GetByID(id);
+            if(user == null)
+                throw new DoesNotExistsException("Such User Does Not Exists!");
+            return user;
+        }
+        public UserCreationDTO GetUser(long fbId) {
+            var user = _userRepository.GetByPredicate(f => f.FbKey == fbId).FirstOrDefault();
+            if (user == null)
+                throw new DoesNotExistsException("Such User Does Not Exists!");
+           return _mapper.Map<UserCreationDTO>(user);
+        }
         public void DeleteImage(string guid) => _service.SendDeleteMessage(guid);
